FIX: Allow "hit by" vehicle message
Hit by vehicle is generic damage which was being skipped by this check.  Only skip out of plugin if damage generic and player not mounted, otherwise proceed and "hit by" code works as intended.  ISSUE: will this fail for any generic damage not caused by player, if there is any?